feat: Implement data management and notification APIs with corresponding Vue components
- Added data management API for backup, cleanup, and database optimization functionalities.
- Introduced notification API for managing user notifications, including fetching, marking as read, and deleting notifications.
- Developed Vue components for data management and notification management, including UI for displaying statistics, backup lists, and notification actions.
- Implemented state management for notifications using Pinia.
- Enhanced user experience with loading states and confirmation dialogs for critical actions.

+++ AI-agent-backend-fastapi/app/models/department.py
@@ -1,14 +1,14 @@
 """部门数据库模型 - 完全按照博客 t_dept 表结构设计"""
-from sqlalchemy import Column, BigInteger, String, Float, DateTime
+from sqlalchemy import Column, Integer, String, Float, DateTime
 from app.core.database import Base
 
 
 class Department(Base):
     """部门表模型 - 对应博客 t_dept 表"""
     __tablename__ = "t_dept"
-    
-    dept_id = Column(BigInteger, primary_key=True, autoincrement=True, comment="部门ID")
-    parent_id = Column(BigInteger, nullable=False, comment="上级部门ID")
+
+    dept_id = Column(Integer, primary_key=True, autoincrement=True, comment="部门ID")  # SQLite使用INTEGER支持自增
+    parent_id = Column(Integer, nullable=False, comment="上级部门ID")  # SQLite使用INTEGER支持自增
     dept_name = Column(String(100), nullable=False, comment="部门名称")
     order_num = Column(Float, nullable=True, comment="排序")
     create_time = Column(DateTime, nullable=True, comment="创建时间")
